TST: Simplify fixture that is equivalent to a function call
