+++ numpy/_core/tests/test_multithreading.py
@@ -1,11 +1,11 @@
 import concurrent.futures
-import string
 import threading
 
 import pytest
 
 import numpy as np
 from numpy._core import _rational_tests
+from numpy._core.tests.string_testing import random_string_list
 from numpy.testing import IS_64BIT, IS_WASM
 from numpy.testing._private.utils import run_threaded
 
@@ -218,16 +218,12 @@ def test_structured_threadsafety2():
     assert arr.dtype is dt
 
 
-def test_stringdtype_multithreaded_access_and_mutation(
-        dtype, random_string_list):
+def test_stringdtype_multithreaded_access_and_mutation(dtype):
     # this test uses an RNG and may crash or cause deadlocks if there is a
     # threading bug
     rng = np.random.default_rng(0x4D3D3D3)
 
-    chars = list(string.ascii_letters + string.digits)
-    chars = np.array(chars, dtype="U1")
-    ret = rng.choice(chars, size=100 * 10, replace=True)
-    random_string_list = ret.view("U100")
+    string_list = random_string_list()
 
     def func(arr):
         rnd = rng.random()
@@ -247,10 +243,10 @@ def test_stringdtype_multithreaded_access_and_mutation(
             else:
                 np.multiply(arr, np.int64(2), out=arr)
         else:
-            arr[:] = random_string_list
+            arr[:] = string_list
 
     with concurrent.futures.ThreadPoolExecutor(max_workers=8) as tpe:
-        arr = np.array(random_string_list, dtype=dtype)
+        arr = np.array(string_list, dtype=dtype)
         futures = [tpe.submit(func, arr) for _ in range(500)]
 
         for f in futures:
